pyproject.toml
[project]
name = "api-consultas-medicas"
version = "1.0.0"
description = "tonysouz"
authors = [
    {name = "tonysouz"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-decouple (>=3.8,<4.0.0)",
    "django-cors-headers (>=4.3.1,<5.0.0)",
    "djangorestframework-simplejwt (>=5.3.1,<6.0.0)",
    "drf-spectacular (>=0.27.1,<0.28.0)",
    "gunicorn (>=21.2.0,<22.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.0,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["apps"]
